Move remaining plot modules down a dir level
