+++ e3sm_diags/driver/area_mean_time_series_driver.py
@@ -13,7 +13,7 @@ from e3sm_diags.driver.utils.io import _write_to_netcdf
 from e3sm_diags.driver.utils.regrid import _apply_land_sea_mask, _subset_on_region
 from e3sm_diags.logger import custom_logger
 from e3sm_diags.metrics.metrics import spatial_avg
-from e3sm_diags.plot.cartopy import area_mean_time_series_plot
+from e3sm_diags.plot import area_mean_time_series_plot
 
 if TYPE_CHECKING:
     from e3sm_diags.parameter.area_mean_time_series_parameter import (
