+++ e3sm_diags/plot/area_mean_time_series_plot.py
