+++ e3sm_diags/plot/taylor_diagram.py
